// DBS-Gateway-BE-001/src/interfaces/application/applicationByIndividual.interface.ts
export interface IEmploymentDetails {
  organizationName: string;
  applicationType: string;
  isApplicationWorkWithAdults: string;
  isApplicationWorkWithChildren: string;
  isThisWorkYourHomeAddress: string;
  applingPositionFor: string;
  isApplicationFreeOfCharge: string;
  isCheckedAddressDetailsWithOriginal: string;
  isCrossCheckedApplicentIdentity: string;
  comment:string;
  isIdentityIndividualBeenVerified: string;
  electronicSignatureImage: string;
}